Extract concurrent create tests
